MRI_Proj/mri_classifier/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.1.0
ultralytics==8.0.196
numpy==1.24.3
opencv-python==4.8.1.78
python-multipart==0.0.6

